Add some quoting for examples
The lack of quoting in some of the examples might make it harder to see what constitutes example and what is punctuation.  Highlighting and fonts in the HTML rendering are subtle and not present in text renderings (any more).

I've also changed some section references.  I'm happy to split that off into a different change or use the `{{SVCB, Section 6.2}}` form if you prefer that.